rrr: convert rpc/server to module (rrr.server)
Merges server.hpp (966 lines) + server.cpp (837 lines) into the
single module unit src/rrr/rpc/server.cpp (~1713 lines). Exports the
Server / ServerConnection / DeferredReply surface plus the rpc
service-binding glue. Sole textual consumer (rrr.hpp umbrella) now
imports rrr.server instead of including the .hpp.
